+++ src/screens/Webtoon.tsx
@@ -8,29 +8,11 @@ import MainBanner from '@/components/Banner/MainBanner';
 import WebtoonList from '@/components/WebtoonList';
 
 import { HEIGHTS, WIDTHS } from '@/styles/dimensions';
-
-interface RouteType {
-  id: number;
-  key: 'new' | 'daily' | 'mon' | 'tue' | 'wed' | 'thu' | 'fri' | 'sat' | 'sun' | 'end';
-  title: string;
-}
-
-const route: RouteType[] = [
-  { id: 0, key: 'new', title: '신작' },
-  { id: 1, key: 'daily', title: '매일' },
-  { id: 2, key: 'mon', title: '월' },
-  { id: 3, key: 'tue', title: '화' },
-  { id: 4, key: 'wed', title: '수' },
-  { id: 5, key: 'thu', title: '목' },
-  { id: 6, key: 'fri', title: '금' },
-  { id: 7, key: 'sat', title: '토' },
-  { id: 8, key: 'sun', title: '일' },
-  { id: 9, key: 'end', title: '완결' },
-];
+import WEBTOON_ROUTE from '@/constants/routes';
 
 export default function HomeScreen() {
   const [index, setIndex] = useState(0);
-  const [routes] = useState(route);
+  const [routes] = useState(WEBTOON_ROUTE);
 
   const scrollY = useRef(new Animated.Value(0)).current;
   const bannerTranslateY = scrollY.interpolate({
